hoang: update

+++ hen_ho/service/src/main/java/com/codegym/service/Impl/ThongBaoServiceImpl.java
@@ -24,7 +24,7 @@ public class ThongBaoServiceImpl implements ThongBaoService {
     }
 
     @Override
-    public ThongBao getThongBaoById(long id) {
+    public ThongBao getThongBaoById(int id) {
         return thongBaoRepository.findById(id).orElse(null);
     }
 }
